+++ src/test/java/com/cgvsu/vectormath/vector/Vector4DTest.java
@@ -1,77 +1,77 @@
 package com.cgvsu.vectormath.vector;
 
-import org.junit.jupiter.api.Test;
-
 import static org.junit.jupiter.api.Assertions.*;
+
+import org.testng.annotations.Test;
 
 public class Vector4DTest {
 
     @Test
     public void testAddition() {
-        Vector4D v1 = new Vector4D(1.0, 2.0, 3.0, 4.0);
-        Vector4D v2 = new Vector4D(4.0, 5.0, 6.0, 7.0);
+        Vector4D v1 = new Vector4D(1.0f, 2.0f, 3.0f, 4.0f);
+        Vector4D v2 = new Vector4D(4.0f, 5.0f, 6.0f, 7.0f);
         Vector4D result = v1.add(v2);
-        assertEquals(5.0, result.get(0), 0.01);
-        assertEquals(7.0, result.get(1), 0.01);
-        assertEquals(9.0, result.get(2), 0.01);
-        assertEquals(11.0, result.get(3), 0.01);
+        assertEquals(5.0f, result.get(0), 0.01);
+        assertEquals(7.0f, result.get(1), 0.01);
+        assertEquals(9.0f, result.get(2), 0.01);
+        assertEquals(11.0f, result.get(3), 0.01);
     }
 
     @Test
     public void testSubtraction() {
-        Vector4D v1 = new Vector4D(4.0, 5.0, 6.0, 7.0);
-        Vector4D v2 = new Vector4D(1.0, 2.0, 3.0, 4.0);
+        Vector4D v1 = new Vector4D(4.0f, 5.0f, 6.0f, 7.0f);
+        Vector4D v2 = new Vector4D(1.0f, 2.0f, 3.0f, 4.0f);
         Vector4D result = v1.subtract(v2);
-        assertEquals(3.0, result.get(0), 0.01);
-        assertEquals(3.0, result.get(1), 0.01);
-        assertEquals(3.0, result.get(2), 0.01);
-        assertEquals(3.0, result.get(3), 0.01);
+        assertEquals(3.0f, result.get(0), 0.01);
+        assertEquals(3.0f, result.get(1), 0.01);
+        assertEquals(3.0f, result.get(2), 0.01);
+        assertEquals(3.0f, result.get(3), 0.01);
     }
 
     @Test
     public void testMultiplication() {
-        Vector4D v1 = new Vector4D(2.0, 3.0, 4.0, 5.0);
-        double scalar = 2.0;
+        Vector4D v1 = new Vector4D(2.0f, 3.0f, 4.0f, 5.0f);
+        float scalar = 2.0f;
         Vector4D result = v1.multiply(scalar);
-        assertEquals(4.0, result.get(0), 0.01);
-        assertEquals(6.0, result.get(1), 0.01);
-        assertEquals(8.0, result.get(2), 0.01);
-        assertEquals(10.0, result.get(3), 0.01);
+        assertEquals(4.0f, result.get(0), 0.01);
+        assertEquals(6.0f, result.get(1), 0.01);
+        assertEquals(8.0f, result.get(2), 0.01);
+        assertEquals(10.0f, result.get(3), 0.01);
     }
 
     @Test
     public void testDivision() {
-        Vector4D v1 = new Vector4D(6.0, 8.0, 10.0, 12.0);
-        double scalar = 2.0;
+        Vector4D v1 = new Vector4D(6.0f, 8.0f, 10.0f, 12.0f);
+        float scalar = 2.0f;
         Vector4D result = v1.divide(scalar);
-        assertEquals(3.0, result.get(0), 0.01);
-        assertEquals(4.0, result.get(1), 0.01);
-        assertEquals(5.0, result.get(2), 0.01);
-        assertEquals(6.0, result.get(3), 0.01);
+        assertEquals(3.0f, result.get(0), 0.01);
+        assertEquals(4.0f, result.get(1), 0.01);
+        assertEquals(5.0f, result.get(2), 0.01);
+        assertEquals(6.0f, result.get(3), 0.01);
     }
 
     @Test
     public void testLength() {
-        Vector4D v = new Vector4D(3.0, 4.0, 5.0, 6.0);
-        double result = v.length();
-        double expected = 9.273618495495704; // Примерное значение
+        Vector4D v = new Vector4D(3.0f, 4.0f, 5.0f, 6.0f);
+        float result = v.length();
+        float expected = 9.273618495495704f; // Примерное значение
         assertEquals(expected, result, 0.00001); // Учитывать погрешность
     }
 
     @Test
     public void testNormalization() {
-        Vector4D v = new Vector4D(1.0, 2.0, 2.0, 2.0);
+        Vector4D v = new Vector4D(1.0f, 2.0f, 2.0f, 2.0f);
         Vector4D result = v.normalize();
-        double length = result.length();
-        assertTrue(Math.abs(length - 1.0) < 0.00001);
+        float length = result.length();
+        assertTrue(Math.abs(length - 1.0f) < 0.00001);
     }
 
     @Test
     public void testDotProduct() {
-        Vector4D v1 = new Vector4D(1.0, 2.0, 3.0, 4.0);
-        Vector4D v2 = new Vector4D(4.0, 5.0, 6.0, 7.0);
-        double result = v1.dotProduct(v2);
-        double expected = 60.0;
+        Vector4D v1 = new Vector4D(1.0f, 2.0f, 3.0f, 4.0f);
+        Vector4D v2 = new Vector4D(4.0f, 5.0f, 6.0f, 7.0f);
+        float result = v1.dotProduct(v2);
+        float expected = 60.0f;
         assertEquals(expected, result, 0.00001); // Учитывать погрешность
 
     }
